requirements.txt
phidata
yfinance
python-dotenv
groq
duckduckgo-search
